feat: improve compose and updated iotsimulator
