+++ IOTSimulator/MQTTTelemetry/app/proguard-rules.pro
