+++ IOTSimulator/MQTTTelemetry/app/src/main/assets/ca.pem
@@ -1,0 +1,34 @@
+-----BEGIN CERTIFICATE-----
+MIIF5TCCA82gAwIBAgIUBUxsCzIHb04SCCvqHZgcWyiUX0cwDQYJKoZIhvcNAQEL
+BQAwgYExCzAJBgNVBAYTAnB0MQ4wDAYDVQQIDAVwb3J0bzEOMAwGA1UEBwwFcG9y
+dG8xDTALBgNVBAoMBHRleWExDDAKBgNVBAsMA2FuYTEMMAoGA1UEAwwDYW5hMScw
+JQYJKoZIhvcNAQkBFhhhbmEuLmxvdXJlaXJvQHNhbHRwYXkuY28wHhcNMjUwMjE3
+MTUwNzMyWhcNMzUwMjE1MTUwNzMyWjCBgTELMAkGA1UEBhMCcHQxDjAMBgNVBAgM
+BXBvcnRvMQ4wDAYDVQQHDAVwb3J0bzENMAsGA1UECgwEdGV5YTEMMAoGA1UECwwD
+YW5hMQwwCgYDVQQDDANhbmExJzAlBgkqhkiG9w0BCQEWGGFuYS4ubG91cmVpcm9A
+c2FsdHBheS5jbzCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCCAgoCggIBAI9eDIyb
+KG9/MbOGUiiT3nHOC5iZp1vcQAWHw6ZG/o46vSu7wzeqUxKWA0votoRt/xaK5UsN
+L+5s2JuSjaHzQpH6ZnLeuM2xBMSutVAAqbF/Q2vaExfT8ReXn4CIjCA16QjwF6YD
+8oVH9Krj3CNLNVXvObzF9G9tbV33jgpdQUCp8NSzFKUx+sNfw7FU3xI+d3X0il/d
+t5nCvx4Rep6ZkYlHOZhhhVJfqw0XCoB/tOEpwET2fM2KZlwLLSRUi0L1KvwEu2rC
++Jd4NAvjSr/aGxUCecv2tgTbDl3qb2PFN5WMTpaoe3gjfNSWSdQoJJFmgwHQwrCf
+7O4t+TbiuYKDkl84yzEleabN8qgLE8xNjKcjUv/uEt2AGOf02HW8ic5YDbcEEQvm
+vtr1rINgZ0ytFRlodWuIRUNC580TZawsyFty+j9LISNys9+LzRT1qz47VAzHfAzT
+gR9FD3qPKko0jcA+TnwDm1Rn0zbMuKOeGJpsTWUbND1SFiVqdfsFqtYrtm3Zu3wS
+eFLRsb9blKIN44HBC+IeVNG8Bfe4FzQd4J85nK5ePx4UMTqgg/o0EGQLutNr/dRJ
+M8T6WM9bogpaO+V9tDPGLZ5JX6dCIf463Y/Fdfp0av2nstm88KG8nCLxEFTTBuYt
+LYmWDuMJ7Ry4cvhwClVOHsntJYsToAnILywfAgMBAAGjUzBRMB0GA1UdDgQWBBTZ
+bYX0vB9RTNb7EeoWxrxXLcUwTTAfBgNVHSMEGDAWgBTZbYX0vB9RTNb7EeoWxrxX
+LcUwTTAPBgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUAA4ICAQBK4kc1UmcU
+fhKy9pAJfB5ffUnJVBDvHOKt1cDZUZg4wS7B6jGzRlg/jCLf1FokjoEZA9lF+de+
+H5ute2Iz3s4koN2uhEyXhPsebAwK2eeSWv+2EEn7Z2RjxomFbeMnf2tBGNdk4Ktl
+7eVP3bYXwi5axTSYx6AIMibWrvmlWXpOwfZA9hYageHhl1aFAx8RXWvRg9cCtpSP
+ZSpPGPrJZtuAK6jv7tdFtgWu+XPW/jqIzC/KaeBaXO8YO0PRYeUXu6MmDGIHg/3o
+oqhrOPZjuJNsec8slb0Tl0mpUPHYHVuGhQu6KcsWToGDObsCSdhWRS11e0VSTrqq
+MdEEeqyuUyijPIITjM7t9zd13uGC5KjPjetWKkBRA0ciy0swTj0GScKsuLLozVLF
+VvA6hxYoFq1RgwTlBq1DX2a9eI7CD5ajYnKmpRA5R5qMsTLixwuKk8quJz8CT196
+9n94FFMZmK1alDVbnaeXfrOCWq+/DB8IhH82/w+mLpyRsPsO0n/jM4S4JYIY75ld
+DY52bFM4NjCaNF/T5xiy3n8O1vlitvMJMSlUeExfKz6y/CcYjSSpMNLcKMarQYWk
+fa9Rxj/meaxt++1QXzniuIeGPtz4Ozv6r+HbL0Qn034bm0uJ9A5MUyGDmMHyKvSQ
+tyJvGMEt1F+j2CRmaDVyXgvle+41+8JXyA==
+-----END CERTIFICATE-----
