+++ IOTSimulator/MQTTTelemetry/app/src/main/res/drawable/ic_launcher_foreground.xml
