+++ IOTSimulator/MQTTTelemetry/app/src/main/res/raw/ca.pem
@@ -1,0 +1,19 @@
+-----BEGIN CERTIFICATE-----
+MIIDCzCCAfOgAwIBAgIUSqAtf9skHtLqofWO/zjHrFdenPswDQYJKoZIhvcNAQEL
+BQAwFTETMBEGA1UEAwwKTXlDdXN0b21DQTAeFw0yNTAyMjAxNTU2MzVaFw0yNjAy
+MjAxNTU2MzVaMBUxEzARBgNVBAMMCk15Q3VzdG9tQ0EwggEiMA0GCSqGSIb3DQEB
+AQUAA4IBDwAwggEKAoIBAQDuW8v6zD3w6dpGWL2XyZ2fwr/c+8/uHfMJoRhV0Hsf
+mUnlqmYSt9BenSeCNycSSBXzLjD4HbDunV31xSlpidTk9DNE4eb4tK9wwRm02T4J
+D/3niHSmLZLh1B3fmZ5gHEOHZcF2+NVq4iuUIe51j+tBmdgQJTxxYxeYzslDCOfh
+kILucStFnZq7uu8+frHV+IUKpreSxWxPEavK9OJabwN17bdPwIKvid9MWWZo5JA4
+I/pjHnCNbYYTh5xQ4Ime6HadiaSG75UZ/6Uiu28Eq9zd7iYiQKmeVNnhNZcv1G2V
+GE6A9ejAQN56U9IwLaGN/ISShE01EMDWZ97PFzfWLVTVAgMBAAGjUzBRMB0GA1Ud
+DgQWBBS5i2S74lHrNNo0ojcK6aBAnKUx8jAfBgNVHSMEGDAWgBS5i2S74lHrNNo0
+ojcK6aBAnKUx8jAPBgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUAA4IBAQDq
++a9DHUpjdY5kPySS6fqsPYuIczNPJBT8IjRLdf+S5RHp04bw/S89Io58w/N7K2PM
+DB7jKWf1t8aVuYQ9e42Ht013DLM6EUPR/tNcAYcVbGf550ZyAoO5c2BrHxB8L1be
+gYH+zqLyjHb4KbrY5JpH010RMkuu9m3ZXP9mBDZSaZmuppr+P42S0lbI7pSIxzOE
+5GQzzvgvNee8Wv7foFvYuVCFTqRtjEmMguh+xmB3YYKYblWGtf/Jiz1Xd0nU2JGB
+RXe71iRWWYgfsh0bBiaom+vTWipr02YZgwK7XUpdWI/GnCEoQwY6fwkjz8lZ+OvC
+yH2idStX6tw4h4mq6SxN
+-----END CERTIFICATE-----
