+++ IOTSimulator/MQTTTelemetry/app/src/main/res/values/colors.xml
