+++ IOTSimulator/MQTTTelemetry/app/src/main/res/values/strings.xml
