+++ IOTSimulator/MQTTTelemetry/app/src/main/res/xml/backup_rules.xml
