+++ IOTSimulator/MQTTTelemetry/gradle/libs.versions.toml
