+++ IOTSimulator/MQTTTelemetry/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Dec 26 14:27:34 WET 2024
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.9-bin.zip
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
