+++ IOTSimulator/MQTTTelemetry/settings.gradle.kts
